+++ Source/embeddings/Builder/deleteDecorator.ts
