+++ Source/embeddings/Builder/index.ts
@@ -7,8 +7,8 @@ export { CompareDecoratedMethods } from './CompareDecoratedMethods';
 export { compare } from './compareDecorator';
 export { EmbeddingAlreadyHasACompareDecorator } from './EmbeddingAlreadyHasACompareDecorator';
 export { EmbeddingAlreadyHasACompareMethod } from './EmbeddingAlreadyHasACompareMethod';
-export { EmbeddingAlreadyHasARemoveDecorator } from './EmbeddingAlreadyHasARemoveDecorator';
-export { EmbeddingAlreadyHasARemoveMethod } from './EmbeddingAlreadyHasARemoveMethod';
+export { EmbeddingAlreadyHasADeletionDecorator } from './EmbeddingAlreadyHasADeletionDecorator';
+export { EmbeddingAlreadyHasADeletionMethod } from './EmbeddingAlreadyHasADeletionMethod';
 export { EmbeddingBuilder } from './EmbeddingBuilder';
 export { EmbeddingBuilderForReadModel } from './EmbeddingBuilderForReadModel';
 export { EmbeddingClassBuilder } from './EmbeddingClassBuilder';
@@ -27,4 +27,4 @@ export { on } from './onDecorator';
 export { ReadModelAlreadyDefinedForEmbedding } from './ReadModelAlreadyDefinedForEmbedding';
 export { RemoveDecoratedMethod } from './RemoveDecoratedMethod';
 export { RemoveDecoratedMethods } from './RemoveDecoratedMethods';
-export { resolveDeletionToEvents } from './removeDecorator';
+export { resolveDeletionToEvents } from './deleteDecorator';
