+++ Source/embeddings/index.ts
@@ -8,8 +8,8 @@ export {
     CompareDecoratedMethods, embedding,
     EmbeddingAlreadyHasACompareDecorator,
     EmbeddingAlreadyHasACompareMethod,
-    EmbeddingAlreadyHasARemoveDecorator,
-    EmbeddingAlreadyHasARemoveMethod,
+    EmbeddingAlreadyHasADeletionDecorator,
+    EmbeddingAlreadyHasADeletionMethod,
     EmbeddingBuilder,
     EmbeddingBuilderForReadModel,
     EmbeddingClassBuilder,
